move static rendering to tableview lib
